GEODE-7258: The function retry logic is modified to handle exception thrown, while trying to connect to a server thats shutdown/closed.
